+++ formik-mui/src/formik/CheckboxGroup.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import CheckboxGroup from '../forms/CheckboxGroup';
 
-const FormikCheckboxGroup = ({fast = true, ...p}) => <CheckboxGroup {...p} Field={require('formik')[fast ? 'FastField' : 'Field']}/>;
+const FormikCheckboxGroup = ({fast, ...p}) => <CheckboxGroup {...p} Field={require('formik')[fast ? 'FastField' : 'Field']}/>;
 export default FormikCheckboxGroup;