+++ formik-mui/src/formik/RadioGroup.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import RadioGroup from '../forms/RadioGroup';
 
-const FormikRadio = ({children, render, fast = true, type, buttons, ...props}) => { // eslint-disable-line no-unused-vars
+const FormikRadio = ({children, render, fast, type, buttons, ...props}) => { // eslint-disable-line no-unused-vars
 	const Field = require('formik')[fast ? 'FastField' : 'Field'];
 	return <Field {...props} component={RadioGroup}>
 		{children}