+++ formik-mui/src/forms/Dropzone/Previews.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import withStyles from '@material-ui/core/styles/withStyles';
-import isImage from '../../utils/isImage';
 import Grid from '@material-ui/core/Grid';
 import DeleteIcon from '@material-ui/icons/Delete';
 import AttachFileIcon from '@material-ui/icons/AttachFile';
@@ -8,12 +7,28 @@ import Fab from '@material-ui/core/Fab';
 import FormHelperText from '@material-ui/core/FormHelperText';
 import grey from '@material-ui/core/colors/grey';
 import LinearProgress from '@material-ui/core/LinearProgress';
+import isImage from '../../utils/isImage';
+import clsx from 'clsx';
 
 const styles = () => ({
-	preview: {
+	allPreviewsContainer: {
 		display: 'flex',
-		alignItems: 'center',
+		flexWrap: 'wrap',
+	},
+	onePreviewContainer: {
+		display: 'flex',
 		flexDirection: 'column',
+		justifyContent: 'stretch',
+		width: 200,
+		padding: 8,
+		wordBreak: 'break-all',
+		position: 'relative',
+		'&:hover $smallPreviewImg': {
+			opacity: 0.3,
+		},
+		'&:hover $removeBtn': {
+			opacity: 1,
+		},
 	},
 	progress: {
 		width: '100%',
@@ -29,8 +44,8 @@ const styles = () => ({
 		transition: '.5s ease',
 		position: 'absolute',
 		opacity: 0,
-		top: -5,
-		right: -5,
+		top: 0,
+		right: 10,
 		width: 40,
 		height: 40,
 		zIndex: 1,
@@ -45,47 +60,55 @@ const styles = () => ({
 		boxShadow: 'rgba(0, 0, 0, 0.12) 0 1px 6px, rgba(0, 0, 0, 0.12) 0 1px 4px',
 		borderRadius: 2,
 	},
-	imageContainer: {
-		position: 'relative',
+	center: {
 		textAlign: 'center',
-		'&:hover $smallPreviewImg': {
-			opacity: 0.3,
-		},
-		'&:hover $removeBtn': {
-			opacity: 1,
-		},
+	},
+	selfCenter: {
+		textAlign: 'center',
 	},
 });
 
 const opacity = file => ({opacity: file.processing ? 0.5 : file.error ? 0.1 : 1});
-const Previews = props => {
-	const {files = [], handleDelete, showFileNames, FormHelperTextProps, classes} = props;
-	const onClick = i => e => {e.preventDefault(); e.stopPropagation(); handleDelete(i);};
+const Previews = ({files = [], handleDelete, showFileNames, FormHelperTextProps = {}, classes, children, cs = {}, ...props}) => {
+	const onClick = i => e => { e.preventDefault(); e.stopPropagation(); handleDelete(i); };
 	if (!files.length) return null;
+
+	FormHelperTextProps.classes = {...FormHelperTextProps.classes, root: clsx(classes.center, (FormHelperTextProps.classes || {}).root)};
 	return (
-		<Grid container style={{padding: '8px'}}>
+		<div className={clsx(classes.allPreviewsContainer, cs.allPreviewsContainer)}>
 			{
 				files.map((file, i) => (
-					<Grid item key={i} className={classes.imageContainer}>
-						<div className={classes.preview}>
-							{(isImage(file) ?
-								<img className={classes.smallPreviewImg} role='presentation' src={file.preview} style={opacity(file)}/>
-								:
-								<AttachFileIcon className={classes.smallPreviewImg} style={opacity(file)}/>
-							)}
-							{file.processing && <LinearProgress classes={{root: classes.progress, colorPrimary: classes.progressColor, barColorPrimary: classes.progressBarColor}}/>}
-							{(file.error || showFileNames) && <FormHelperText {...FormHelperTextProps} error={Boolean(file.error)}>{file.error || file.name}</FormHelperText>}
-						</div>
+					<div key={i} className={clsx(classes.onePreviewContainer, cs.onePreviewContainer)}>
+						<Grid item xs={12} className={classes.selfCenter}>
+							<a target='_blank' aria-label={file.name} href={file.path}>
+								{(isImage(file) ?
+									<img className={clsx(classes.smallPreviewImg, cs.smallPreviewImg)} role='presentation' src={file.preview} style={opacity(file)}/>
+									:
+									<AttachFileIcon className={clsx(classes.smallPreviewImg, cs.smallPreviewImg)} style={opacity(file)}/>
+								)}
+							</a>
+						</Grid>
+						{file.processing && <LinearProgress classes={{root: classes.progress, colorPrimary: classes.progressColor, barColorPrimary: classes.progressBarColor}}/>}
+						{(file.error || showFileNames) && <Grid item xs={12}>
+							<a target='_blank' aria-label={file.name} href={file.path}>
+								<FormHelperText {...FormHelperTextProps} error={Boolean(file.error)}>{file.error || file.name}</FormHelperText>
+							</a>
+						</Grid>}
+						{file.uploaded && <Grid item xs={12}>
+							{React.Children.map(children, (child, j) => React.cloneElement(child, {...props, key: i + '' + j, file, index: i}))}
+						</Grid>}
 						<Fab onClick={onClick(i)}
 							aria-label='Delete'
 							className={classes.removeBtn}>
 							<DeleteIcon/>
 						</Fab>
-					</Grid>
+					</div>
 				))
 			}
-		</Grid>
+		</div>
 	);
 };
 
-export default withStyles(styles)(Previews);
+const Previews1 = withStyles(styles)(Previews);
+export default ({classes, ...props}) => <Previews1 cs={classes} {...props}/>;
+