+++ formik-mui/src/forms/formikToMuiProps.js
@@ -1,4 +1,5 @@
 import {getIn} from 'formik';
+import isDate from '../utils/isDate';
 
 export default function({
 	field = {},
@@ -13,8 +14,8 @@ export default function({
 	const {errors = {}, touched = {}, isSubmitting, dirty} = form;
 
 	const fErr = name && getIn(errors, name);
-	window.getIn = getIn;
-	const fieldError = (dirty || (name && getIn(touched, name))) && typeof fErr === 'string' ? fErr : null;
+	const fieldTouched = (name && getIn(touched, name));
+	const fieldError = (dirty || fieldTouched) && typeof fErr === 'string' ? fErr : null;
 
 	const extraProps = {};
 
@@ -32,6 +33,10 @@ export default function({
 				field.value = typeof field.value === 'undefined' ? [] : field.value;
 				break;
 
+			case 'date':
+				field.value = typeof field.value === 'undefined' ? '' : isDate(field.value) ? field.value.toISOString().split('T')[0] : field.value;
+				break;
+
 			default: typeof field.value === 'undefined' ? '' : field.value;
 		}
 	}
@@ -46,6 +51,7 @@ export default function({
 		...props,
 		...field,
 		...extraProps,
+		// touched: fieldTouched,
 		error: error || Boolean(fieldError),
 		helperText: fieldError || props.helperText,
 	};