+++ formik-mui/CHANGELOG.md
@@ -3,6 +3,213 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.20](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.19...@settlin/formik-mui@1.3.20) (2019-04-02)
+
+
+### Bug Fixes
+
+* multiple fixes ([b4869a2](http://github.com/settlin/node-monorepo/formik-mui/commit/b4869a2))
+
+
+
+
+
+## [1.3.19](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.18...@settlin/formik-mui@1.3.19) (2019-03-30)
+
+
+### Bug Fixes
+
+* filterfield without formik ([3f482f7](http://github.com/settlin/node-monorepo/formik-mui/commit/3f482f7))
+
+
+
+
+
+## [1.3.18](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.17...@settlin/formik-mui@1.3.18) (2019-03-25)
+
+**Note:** Version bump only for package @settlin/formik-mui
+
+
+
+
+
+## [1.3.17](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.16...@settlin/formik-mui@1.3.17) (2019-03-23)
+
+
+### Bug Fixes
+
+* dropzone & select ([54ffb4a](http://github.com/settlin/node-monorepo/formik-mui/commit/54ffb4a))
+
+
+
+
+
+## [1.3.16](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.15...@settlin/formik-mui@1.3.16) (2019-03-23)
+
+
+### Bug Fixes
+
+* dropzone & select ([5c38d1f](http://github.com/settlin/node-monorepo/formik-mui/commit/5c38d1f))
+
+
+
+
+
+## [1.3.15](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.14...@settlin/formik-mui@1.3.15) (2019-03-23)
+
+
+### Bug Fixes
+
+* dropzone & select ([2ddfa49](http://github.com/settlin/node-monorepo/formik-mui/commit/2ddfa49))
+
+
+
+
+
+## [1.3.14](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.13...@settlin/formik-mui@1.3.14) (2019-03-23)
+
+
+### Bug Fixes
+
+* dropzone & select ([a8499a5](http://github.com/settlin/node-monorepo/formik-mui/commit/a8499a5))
+
+
+
+
+
+## [1.3.13](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.12...@settlin/formik-mui@1.3.13) (2019-03-23)
+
+
+### Bug Fixes
+
+* dropzone & select ([fc0272c](http://github.com/settlin/node-monorepo/formik-mui/commit/fc0272c))
+
+
+
+
+
+## [1.3.12](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.11...@settlin/formik-mui@1.3.12) (2019-03-22)
+
+
+### Bug Fixes
+
+* dropzone ([66bcda9](http://github.com/settlin/node-monorepo/formik-mui/commit/66bcda9))
+
+
+
+
+
+## [1.3.11](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.10...@settlin/formik-mui@1.3.11) (2019-03-22)
+
+
+### Bug Fixes
+
+* dropzone ([7032549](http://github.com/settlin/node-monorepo/formik-mui/commit/7032549))
+
+
+
+
+
+## [1.3.10](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.9...@settlin/formik-mui@1.3.10) (2019-03-22)
+
+
+### Bug Fixes
+
+* dropzone ([7f0142b](http://github.com/settlin/node-monorepo/formik-mui/commit/7f0142b))
+
+
+
+
+
+## [1.3.9](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.8...@settlin/formik-mui@1.3.9) (2019-03-22)
+
+
+### Bug Fixes
+
+* dropzone ([aa925ea](http://github.com/settlin/node-monorepo/formik-mui/commit/aa925ea))
+
+
+
+
+
+## [1.3.8](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.7...@settlin/formik-mui@1.3.8) (2019-03-22)
+
+
+### Bug Fixes
+
+* select mui ([911e69c](http://github.com/settlin/node-monorepo/formik-mui/commit/911e69c))
+
+
+
+
+
+## [1.3.7](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.6...@settlin/formik-mui@1.3.7) (2019-03-22)
+
+
+### Bug Fixes
+
+* dropzone ([9405d2b](http://github.com/settlin/node-monorepo/formik-mui/commit/9405d2b))
+
+
+
+
+
+## [1.3.6](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.5...@settlin/formik-mui@1.3.6) (2019-03-22)
+
+
+### Bug Fixes
+
+* dropzone ([fa8713f](http://github.com/settlin/node-monorepo/formik-mui/commit/fa8713f))
+
+
+
+
+
+## [1.3.5](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.4...@settlin/formik-mui@1.3.5) (2019-03-22)
+
+
+### Bug Fixes
+
+* dropzone ([67fb4a5](http://github.com/settlin/node-monorepo/formik-mui/commit/67fb4a5))
+* dropzone ([dbaa2f7](http://github.com/settlin/node-monorepo/formik-mui/commit/dbaa2f7))
+
+
+
+
+
+## [1.3.4](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.3...@settlin/formik-mui@1.3.4) (2019-03-22)
+
+
+### Bug Fixes
+
+* dropzone ([29c0881](http://github.com/settlin/node-monorepo/formik-mui/commit/29c0881))
+
+
+
+
+
+## [1.3.3](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.2...@settlin/formik-mui@1.3.3) (2019-03-20)
+
+
+### Bug Fixes
+
+* disabled/readOnly in react-select, remove mui className ([1b9a011](http://github.com/settlin/node-monorepo/formik-mui/commit/1b9a011))
+
+
+
+
+
+## [1.3.2](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.1...@settlin/formik-mui@1.3.2) (2019-03-18)
+
+
+### Bug Fixes
+
+* check name while setting error ([2cca678](http://github.com/settlin/node-monorepo/formik-mui/commit/2cca678))
+
+
+
+
+
 ## [1.3.1](http://github.com/settlin/node-monorepo/formik-mui/compare/@settlin/formik-mui@1.3.0...@settlin/formik-mui@1.3.1) (2019-03-16)